+++ TTC/controllers/ResponsesController.js
@@ -37,16 +37,16 @@ app.controller("ResponsesController", function($scope, $http) {
 
     //Delete an user response
     $scope.deleteResponse = function (Username) {
+        if (confirm("Are you sure you want to delete this user query?")) {
+            var delResponse = Username;
 
-        var delResponse = Username;
-
-        $http.delete("http://localhost:8086/responses/" + delResponse).then(function (response) {
-            alert("Response Deleted Successfully!");
-            refreshResponses();
-        }, function (response) {
-            alert("Response Deleting Failed !");
-        });
-
+            $http.delete("http://localhost:8086/responses/" + delResponse).then(function (response) {
+                alert("Response Deleted Successfully!");
+                refreshResponses();
+            }, function (response) {
+                alert("Response Deleting Failed !");
+            });
+        }
     };
 
 });
